PET-173 Refactor: 도메인 패키지 생성
