+++ src/main/java/org/retriever/server/dailypet/domain/user/User.java
@@ -1,4 +1,4 @@
-package org.retriever.server.dailypet.user;
+package org.retriever.server.dailypet.domain.user;
 
 import lombok.Getter;
