+++ src/main/java/org/retriever/server/dailypet/domain/user/UserController.java
@@ -1,4 +1,4 @@
-package org.retriever.server.dailypet.user;
+package org.retriever.server.dailypet.domain.user;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
